Move new history store into place
