+++ client/src/store/historyStore/index.js
@@ -2,3 +2,4 @@ export { collectionElementsStore } from "./collectionElementsStore";
 export { datasetStore } from "./datasetStore";
 export { historyItemsStore } from "./historyItemsStore";
 export { historyChangedItemsStore } from "./historyChangedItemsStore";
+export { historyStore } from "./historyStore";
